Move SegmentExtractingNoder to noding package
